Refactor AVR scheduler wrapper into dedicated library
